added IUmbracoSearcher and ExamineSearcher
